Raw code port to TensorFlow
This changes the PyTorch implementation of the GAD algorithm into a TensorFlow version.
This is a direct translation of the source code with no optimization whatsoever.

The main change required is that TensorFlow forbids assigning to tensors.
During the iterative diffusion process, we build new Tensors using tf.stack. This requires some trick with padding since the usual indexing from PyTorch does not work.